+++ src/transfer_function/mod.rs
@@ -423,14 +423,38 @@ impl<T: Clone, U: Time> TfGen<T, U> {
     /// use automatica::{poly, Tf};
     /// use num_complex::Complex as C;
     /// let tf = Tf::new(poly!(1., 2., 3.), poly!(-4., -3., 1.));
-    /// assert_eq!(-8.5, tf.eval(3.));
-    /// assert_eq!(C::new(0.64, -0.98), tf.eval(C::new(0., 2.0_f32)));
+    /// assert_eq!(-8.5, tf.eval_by_val(3.));
+    /// assert_eq!(C::new(0.64, -0.98), tf.eval_by_val(C::new(0., 2.0_f32)));
     /// ```
-    pub fn eval<N>(&self, s: N) -> N
+    pub fn eval_by_val<N>(&self, s: N) -> N
     where
         N: Add<T, Output = N> + Clone + Div<Output = N> + Mul<Output = N> + Zero,
     {
-        self.num.eval(s.clone()) / self.den.eval(s)
+        self.num.eval_by_val(s.clone()) / self.den.eval_by_val(s)
+    }
+}
+
+impl<T, U: Time> TfGen<T, U> {
+    /// Evaluate the transfer function.
+    ///
+    /// # Arguments
+    ///
+    /// * `s` - Value at which the transfer function is evaluated.
+    ///
+    /// # Example
+    /// ```
+    /// use automatica::{poly, Tf};
+    /// use num_complex::Complex as C;
+    /// let tf = Tf::new(poly!(1., 2., 3.), poly!(-4., -3., 1.));
+    /// assert_eq!(-8.5, tf.eval(&3.));
+    /// assert_eq!(C::new(0.64, -0.98), tf.eval(&C::new(0., 2.0_f32)));
+    /// ```
+    pub fn eval<'a, N>(&'a self, s: &'a N) -> N
+    where
+        T: 'a,
+        N: 'a + Add<&'a T, Output = N> + Div<Output = N> + Mul<&'a N, Output = N> + Zero,
+    {
+        self.num.eval(s) / self.den.eval(s)
     }
 }
 
@@ -451,6 +475,7 @@ impl<T: Display + One + PartialEq + Signed + Zero, U: Time> Display for TfGen<T,
 mod tests {
     use super::*;
     use crate::{poly, Continuous, Discrete};
+    use num_complex::Complex;
 
     #[test]
     fn transfer_function_creation() {
@@ -459,6 +484,22 @@ mod tests {
         let tf = TfGen::<_, Continuous>::new(num.clone(), den.clone());
         assert_eq!(&num, tf.num());
         assert_eq!(&den, tf.den());
+    }
+
+    #[test]
+    fn evaluation() {
+        let tf = TfGen::<_, Continuous>::new(poly!(-0.75, 0.25), poly!(0.75, 0.75, 1.));
+        let res = tf.eval(&Complex::new(0., 0.9));
+        assert_abs_diff_eq!(0.429, res.re, epsilon = 0.001);
+        assert_abs_diff_eq!(1.073, res.im, epsilon = 0.001);
+    }
+
+    #[test]
+    fn evaluation_by_value() {
+        let tf = TfGen::<_, Continuous>::new(poly!(-0.75, 0.25), poly!(0.75, 0.75, 1.));
+        let res1 = tf.eval(&Complex::new(0., 0.9));
+        let res2 = tf.eval_by_val(Complex::new(0., 0.9));
+        assert_eq!(res1, res2);
     }
 
     #[test]